Removed the older version of the magnetic field class.
Changed the basic data type of the magnetic field from double to float.
This should be enough, and can be changed by modifying a single typedef
in the header file if necessary (how can we do this at runtime?).  The
associated increase in speed is about a factor two for initialization,
but since double brings me heavily in swapping mode I can't comment on
the speed improvement for the actual interpolation.

I added the radial and phi components of the field.  They are the more
natural components and we could consider tracking in fixed phi planes
to speed up calculation (maybe only as first order approximation).  The
components are calculated when the field map is read in.  The increased
memory use should offset the gains in speed when this is used.  I also
added the methods to retrieve the cartesian field components and the
cylindrical field components.  GetFieldValue() still returns the cartesian
values.

I benchmarked the trilinear versus nearest-neighbor interpolation with the
following results:
  Initialization: 3 sec and 285067 usec
  Reading field map: 97 sec and 579924 usec
  Interpolation method set to 0 (kTrilinearInterpolation)
  Sampling (total of 100000 samples): 0 sec and 331779 usec
  Sampling (per sample): 3.31779 usec
  Interpolation method set to 2 (kNearestNeighborInterpolation)
  Sampling (total of 100000 samples): 0 sec and 170973 usec
  Sampling (per sample): 1.70973 usec
This is wall time, not CPU time, but the relative numbers should be
reliable.  The difference in field value is in the 1% range (for some
handpicked points).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@558 b9700a98-9937-0410-b777-f318d6eb5bff