Implemented a uniform Index() function that will map the three-dimensional grid coordinates to the linear ones (with the prospect of putting phi as a more significant index, and z as the least significant index).  Right now it still keeps the old 3-to-1 dimensional mapping.  I also reduced the number of operations in getting a field value (mostly in determining the grid cell you are in), but reordering and using modf only once.
If you have boost::iostreams, you can now gzip your field map and read it
in that way.  It doesn't change timing or memory usage, but at least it
doesn't take up 600Mb anymore (I am always out of hd space).  To use this,
re-run make config with the option ADD=-D__USE_IOSTREAMS and make sure
that the right libraries are added in the boost link list of GNUmakefile.
It works with boost 1.41, and I haven't tested this anywhere else, but it
is not activated by default.  If we pull in boost::iostreams as a QwAnalysis
dependency, this can be activated by default (or it can check the file
extension).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@559 b9700a98-9937-0410-b777-f318d6eb5bff