Modifications to the options processing.  All VQwSubsystems inherit a static void DefineOptions() that can be used to define options on gQwOptions using AddOptions()().  Unfortunately the C++ ISO standard does not require virtual static methods.  This was apparently an oversight in the drafting of the standard, but no-one seemed to be bothered by it much.
So we have to apply some 'tricks' (anti-global-warming folks take note)
to get the desired behavior.  If you want to add your options to the
recognized list, you have to add a call to your static function in the
respective QwOptionsTracking.h or QwOptionsParity.h headers where the
QwOptions::DefineOptions method is implemented.  Take a look at QwSim-
Tracking.cc for details.  This maintains the separation of the parity
and tracking libraries.

If virtual static methods existed we could have done this in a cleaner
way using the following structure:
class QwOptionsTracking: public QwOptions {
 virtual static void DefineOptions() {
  /* define tracking options */
 }
}
extern QwOptionsTracking gQwOptions;


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@562 b9700a98-9937-0410-b777-f318d6eb5bff