Changed QwDetectorID::fRegion from Int_t to EQwRegionID and modified all other occurrences (mostly in QwHit, VDC and HDC class).
Removed some unused but transferred options from a couple of functions
in QwTrackingTreeCombine.

Using QwLog in QwTrackingWorker and QwTrackingTreeCombine now.

Renamed TsSearch to SearchTreeLines in QwTrackingTreeSearch.  This class
should be merged with QwTrackingTreeRegion and QwTrackingTree.  They are
so similar in scope that I wonder why we have to pass them to each other
all the time.

Inherit VQwTrackingElement by QwTrack, QwPartialTrack, QwTrackingTreeLine.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@568 b9700a98-9937-0410-b777-f318d6eb5bff