Implementation of the QwHitPattern class that contains the hit patterns, which used to be in a set of arrays.  This makes for more portable code, is more readable in QwTrackingWorker (e.g. pattern.AddHitList(hitlist)), and will be passed to the tree for treeline finding.  My plan is to com- bine QwTrackingTree, QwTrackingTreeRegion, and QwTrackingTreeSearch in one class QwTrackingTree.  An object of that class would then take the QwHitPattern as argument and return tree lines much more naturally than how it is done now.
At the moment only region 2 uses QwHitPattern, but region 3 will be easy
to implement too.  It might already work to just replace the respective
calls in QwTrackingWorker, but I haven't tested it yet.

The main benefit with this commit is that the dependency on the deprecated
rcDETRegion global geometry list was again slightly reduced.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@572 b9700a98-9937-0410-b777-f318d6eb5bff