Putting the ROOT::Math::Interpolator stuff behind #ifdef guards.
This is not untested code, but due to its nature it might require
another modification before it works perfectly on all systems.  So
give it a day or two if you don't want to risk it.

The GNUmakefile was modified to provide the root-config --features
list in the form of -D__ROOT_HAS_FEATURE defines.  They can then be
used to check for the presence of a feature in the headers and source.
The magic is s/\(\s*\)\([a-zA-Z0-9_]*\)/\1-D__ROOT_HAS_\2/g ;-)
(smiley not included)

root-config was already a requirement.  It has of course also seen
a number of changes over time, but I believe that --features has been
there for a long enough time now.  Let me know if you run into trouble.

E.g. in QwTrajectory:
  #ifdef __ROOT_HAS_MATHMORE
  # include <Math/Interpolator.h>
  #else
  # warning "The QwTrajectory look-up table momentum determination will not be built!"
  #endif

Similarly, the ROOT_VERSION macro is used to determine the API for
the Interpolator code (and this is also guarded by __ROOT_HAS_MATHMORE).
  #if ROOT_VERSION_CODE >= ROOT_VERSION(5,22,0)
    // Newer version of the MathMore plugin use kPOLYNOMIAL
    #warning "Using newer version of MathMore plugin."
    ROOT::Math::Interpolator inter(size, ROOT::Math::Interpolation::kPOLYNOMIAL);
  #else
    // Older version of the MathMore plugin use kPOLYNOMIAL
    #warning "Using older version of MathMore plugin."
    ROOT::Math::Interpolator inter(size, ROOT::Math::Interpolation::POLYNOMIAL);
  #endif

You probably will need to do a make config before this works.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@580 b9700a98-9937-0410-b777-f318d6eb5bff