Modifications to the Main detector classes and added in QwCombinedPMT class to process the combination of current mode PMTs for main detector. The main detector PMTs can be combined by specify the combinations in the input configuration file qweak_adc.map. The combination can be very flexible as a combination of any two, three or more PMTs.
The example in qweak_adc.map instantiates 30 PMT channels, which correspond to:

   1. 18 individual main detector PMTs (including the fully assembled background detector)
   2. 8 combinations of 2 diagonally opposed PMTs and a combination of all PMTs
   3. 3 additional diagnostic detector channels 

A new branch in the output tree was implemented, corresponding to each of the 30 PMT channels. The names in the ttree automatically follow the detector name defination in the configuration file.

An example output root file can be generated by run the analyzer qwanalysis_adc on a mock data file (generated through the qwmockdatagenerator) for scripting and/or GUI tools development and secondary processing. 


Peiqing


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@593 b9700a98-9937-0410-b777-f318d6eb5bff