Added support for --online flag to the QwTracking executable.  This was the last bit that still needed the class QwCommandLine.  What was holding this up was support for specifying just --online instead of --online true (for a boolean-valued option).  This is now solved by using zero_tokens() in boost.
Also, added a single line Clear() to the constructor of the QwHitRootContainer
which seems necessary in some cases though not in our executables.  In
particular, when moving around hits in the QwHitRootContainer, the static
list behind everything seems to get confused.  Clearing the class member
reference to that list seems to be necessary in those cases.

For information purposes: QwHitRootContainer does the memory allocation
ahead of time by creating a large (enough) list of hit pointer.  This speeds up
the single events because no memory has to be allocated anymore; only pointers
have to be copied.  When a new QwHitRootContainer objects is created, the old
list of hit pointers is reused and this can cause old hits to reappear in the
new object.  The Clear() operation prevents this without slowing things down.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@598 b9700a98-9937-0410-b777-f318d6eb5bff