Moved some global constant #defines to their respective headers, not in the globals.h header.
The treesearch is now a standalone function call, not a combination of
BeginSearch(), SearchTreeLines(), and EndSearch().

Region 3 tree lines are stored in the ROOT output along with their detector
identification (not yet QwDetectorInfo pointer, just region/package/direction).

Allow for more missed VDC wires per set of 8 (modifiable by #define in tree
search header).  The region cosmic events have usually less than 5 wire hits,
and often two of those are reconstructed with a drift distance that puts the
hit outside of the chamber (discarded).  In practice (for run 1566) this means
that only 3 wires per event are usable.  With the new allowance for missed
wires a few more events have a reconstructed tree line.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@601 b9700a98-9937-0410-b777-f318d6eb5bff