There were still problems getting good partial tracks from the region 3 VDCs, and I always thought that this was because of the tracking stage. Turns out that it might be related to a bug in the Geant4 Monte Carlo... This commit might fix the issue (it works on at least a few more tracks).
The Geant4 Monte Carlo trees are used as input (QwSim_100.root) with the
help of the QwTreeEventBuffer.  That class takes the position and momentum
of the track at the four VDC wire planes and 'manually' creates about 5
hits for each track based on the angle between the plane and the momentum.
The momentum is taken from the Monte Carlo tree.

The reference frames used in the Monte Carlo are different from the frames
used in the analysis code (probably with good reason, although I am not
always sure).  The momentum in the Monte Carlo file is stored in a global
reference frame (z downstream, y up) and in a local reference frame (xy in
the VDC wire plane, z upstream, local x equal to global x).  Somewhere in
the Geant4 internals the conversion of the global momentum to the local
momentum components fails, leading to a local momentum of > 4 GeV.

  // The global momentum components are used because there seems to be a problem
  // with the local components in the MC file.  For one event the following
  // values are stored (wdc, 2009-12-31, event 0 in QwSim_100.root):
  //         local     global (units: cm and MeV)
  //       x =  6.80    6.80
  //       y = -7.65    269.7
  //       z = -0.00    439.6
  //      Px = 12.13    12.13
  //      Py = -3498    441.3   <-- too large in local frame
  //      Pz = -2113    1067.3  <-- too large in local frame
  //     |P| = 4086.5   1155.0
  // (for a beam energy of 1165.0 MeV)

The slopes of the track with the wire plane are calculated wrong, therefore
the wrong wires are activated.  Those wrong wires do lead to a nicely recon-
structed tree line, but the two tree lines in the front and back VDC planes
do not line up and are thus discarded.

This commit changes the QwTreeEventBuffer to use the global momentum vector
instead of the local vector.  This requires an additional transformation in
the code.  This also fixes the problem where we used to have too many wires
that were activated in each plane, so now we don't need an arbitrary cut-off
anymore and we can use the edge of the chamber.


Other changes:
- QwDetectorInfo now stores the sin and cos of the angle too.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@602 b9700a98-9937-0410-b777-f318d6eb5bff