The QwEvent class should have received some more thought before we/I implemented it (a trivial use case would have caught the following).
When trying to do a simple MC generated versus reconstructed analysis,
problems immediately occur due to the static list of tracking objects
in QwEvent: the generated and reconstructed events contain the same
information :-\  This static list is there to avoid expensive new and
delete calls, reducing complexity from O(n^2) to O(n).  There is no
real reason to keep the list static, it can also be class local.  Then
the event should not be deleted but rather Clear()'ed or Reset().  The
other option is to use STL vectors (which now seems to work, the trick
was to use an std::vector<TObject*>, not a std::vector<TObject>, ROOT
handles the saving of objects pointed to transparently*).

This commit changes the QwHit list in QwEvent to a user #define-able
strategy, defaulting to the STL vectors.  In the future this will be
profiled and the fastest will be chosen since they all have the same
output in ROOT tree format.  Soon, I will change the tree lines and
partial tracks in QwEvent to the same scheme, but again, that should
not disrupt anyone but myself.

(*) QwHitContainer is now defined as std::vector<QwHit>, which is why
it doesn't write to ROOT files well.  If we were to define the QwHit-
Container as an std::vector<QwHit*>, we would be able to write to a
ROOT tree easily (not tested).  QwHitContainer would maintain owner-
ship of the QwHits and delete them on destruction.  Just an idea to do
away with the QwRootHitContainer conversions...


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@610 b9700a98-9937-0410-b777-f318d6eb5bff