Added branch for the restructuring of the momentum/Q2 determination routines.
Right now the QwTrajectory, QwBridge, QwMagneticField, etc, are all single
classes without inheritance structure.  In this branch, we will try out an
inheritance scheme that should allow us to have a uniform interface to various
momentum determination schemes.

The design idea right now is to have the following base class:

  class VQwMomentumDetermination;
  /* Defines the interface to all momentum determination objects
   * Provides virtual functions such as:
   *  - QwTrack* Calculate (QwPartialTrack* front, QwPartialTrack* back)
   *    calculate a single track based on two partial tracks
   *  - QwTrack* Calculate (QwPartialTrack* front, QwPartialTrack* back, QwTrack* guess)
   *    calculate a single track based on two partial tracks and a first guess
   *  - std::vector<QwTrack*> Calculate (QwPartialTrack* front, QwPartialTrack* back)
   *    calculate a vector of possible tracks based on two partial tracks
   *  - std::vector<QwTrack*> Calculate (std::vector<QwPartialTrack*> front, std::vector<QwPartialTrack*> back)
   *    calculate a vector of possible tracks based on same-length vectors of partial tracks
   */

The following three derived tracking methods should be supported (with as little
as possible 'abuse' of the base class):
- direct RK{1234} shooting from front to back (given guess)
- forced bridging at magnet center given front and back
- lookup in momentum table given front and back

Now that I am typing this, I realize that there are still a lot of class design
decisions that need to be made...



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/restruct-Q2-det@631 b9700a98-9937-0410-b777-f318d6eb5bff