Many modifications related to the Compton analyzer.  Although functionally there is not that much improvement in the analyzer, this now runs smoothly on the EEL126 data files for the photon detector and produces output that is comparable in scope as the 'primitive' CMU analyzer.  It also pointed at a number of issues in the CODA crl files...
A modification in the VQwSubsystem allows to add 'expert trees' in the ROOT
output which do not have to follow the format of the standard parity ROOT
trees MPS_Tree and Hel_Tree.  The use of this expert tree is discouraged if
you can avoid it!  Put everything in MPS_Tree and Hel_Tree where it has a
uniform interface and can be easily analyzed.  If you need an expert tree,
you probably know it already and you can figure out how to use the routines.
The tree ComptonPhoton (in TDirectory 'expert') only contains the number of
sampling ADC events per helicity window, but that will be expanded.

There is support for average waveform determination.  At this point only per
helicity period averages are calculated, see below for QwHelicity comments
which prevent this from being more useful.  Also, I need to look into all
the work on running sums that has been done already.  This should be similar.

Small changes:
- Readout error are caught by double checking what we expect to read from the
  buffer and what we actually do read.
- Time windowed integration on the sample data is now possible.  For now only
  used in calculating the pedestal based on pre-trigger events.

Remaining issues that are somewhat important:
- QwHelicity: since there is no real helicity signal yet, until we hook stuff
  up to the helicity line in the counting house, I am trying to 'fake' that
  signal by giving it a seed and proper initialization.  This is similar to
  what the mock data generator does to write the helicity to the CODA file.
  It isn't activated in the QwCompton::main() routine because it doesn't work
  as well as it should (too many predictor resets for some reason).
- Template support for SIS3320 sampling data type: right now everything is one
  data type (typedef'ed as MQwSIS3320_Type for consistency).  The raw data is
  unsigned short (12-bit), but that is too short even for helicity window
  accumulated samples.  And for running sums and averages we would probably
  need double anyway.  Ideally, I would have liked
    template<class T> class MQwSIS3320_Samples;
  but it's a bit too late to retrofit and inline all the functions.  So what
  will probably happen is that we will just keep on using a typedef Double_t.
- QwComptonElectronDetector: there is a meeting scheduled to discuss who/what
  will be done about this.  The CODA data stream format has been finalized
  and we can now think about reading it in.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@639 b9700a98-9937-0410-b777-f318d6eb5bff