Modifications to the tree line matching between the two like-pitched VDC planes in region 3.  There were problems with the slope calculation of the final tree line.  Now the slope of the tree line after combining the two individual tree lines in the two VDC planes is correct (pen-and-paper cross check).
One thing to keep in mind when interpreting these slopes: the individual tree
lines have slopes defined as the inverse of the slope of the combined tree line.
This will change to make slopes consistent between region 2 and region 3 (you
can define the slope with respect to the wire plane, or with respect to the
normal on the plane).

Smaller modifications:
- QwHit: warn when using some getters on deprecated variable interpretations,
  only relevant to tracking really. Also output the ambiguity of a VDC hit by
  putting a (?) behind the hit parameters.
- nodenode and treenode: removed some remaining debugging output.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@647 b9700a98-9937-0410-b777-f318d6eb5bff