Added examples of what should be shown for the luminosity monitors.  There is a button to change how one views the output of the lumis.  "User" is for everyone that takes shifts.  Detail is more for experts (I guess).
In the user view there are two 2D histograms.  The first is Asymmetry vs Lumi #.The second is Asymmetry width vs lumi #.  Also in the width histogram is a line that will eventually scale with current to show the expected width of the lumis due to counting statistics compared to what is being measured. This will visualize how much the target is boiling.

In the detail view there are 8 histograms.  These have the asymmetry distributions of all the luminosity monitors.  I would like to have it set up so that the top histo is the lumi that is on the top of the beam line etc.  The middle histogram slot is left empty to be a place holder for the "beam".

I haven't figured out how to open a data file with the gui nor do I know how to get the data from the data file into a histogram. That will be my next project.




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@650 b9700a98-9937-0410-b777-f318d6eb5bff