After the discussion about the GEM data at the analysis workshop on Friday, I modified the routines that generate QwHit lists from the Monte Carlo data to generate multiple QwHits for each track going through a GEM.  An additional function QwTreeEventBuffer::CreateHitRegion1(detectorinfo,x,y) takes care of this.  It is similar ot the CreateHitRegion3() function, and could easily be expanded for BONuS charge-per-strip mode, by just taking Q = f(1/distance_r3).
I added two GEM cluster finding classes (still mostly empty):
- QwGEMCluster will contain the cluster information (x and delta x; for x =
  r, phi) as well as references back to the hits used to generate the cluster
- QwGEMClusterFinder will implement the cluster finding itself, at this point
  a simple average of the strips (within a reasonable range).  The result of 
  the cluster finding will be a std::vector<QwGEMCluster>.                   

The tracking geometry file (qweak_new.geo) contained already lines for the
GEMs, but it is not clear how correct they are (especially since a comment
there indicates that the actual positions and the MC positions are different
by 15 cm in z...).  Also, I added an arbitrary number of strips, since that
is needed for the position-to-strip conversion.  Other detector lines have
the active width given in x and y coordinates, but here it might be better
to give the width as r and phi extents.  The central position can still be
in x and y.  Distance between phi strips (which are radial) in particular
should be defined uniquely as there can be some confusion there...

QwHit does not have fPhiPosition and fRPosition fields anymore, and the <<
operator was modified.  A couple of accessors in QwHit.cc that assumed that
detector position and resolution were stored in the QwHit were removed (and
all locations with this deprecated use pattern were modified to use the info
pointer, mostly in older tracking code).  This makes QwHit.cc again QwLog-free,
and fixes loading the simple shared libraries in ROOT.

Smaller modifications:
- Number of allowed missed wires per VDC plane group of 8 is set back to 3.
  It was at 7 for a while for debugging purposes.  3 missed per 8 wires is
  a reasonable assumption, at least more resonable than 7 missed for 8 wires...
- Some operator<< were modified to have an explicit const argument, otherwise
  we get warnings with the templated QwLog::operator<< that expects a const.
  Output stream operators should not modify their contents anyway.  This is
  applicable to QwDetectorInfo and QwHitContainer.
- The treelines have now more meaningful names in region 3: fR3Offset is the
  first wire of the demuxed group of 8 that is under consideration.  In that
  group of 8, fR3FirstWire and fR3LastWire are between 0 and 7.
- The tree search in R3 was initiated for all 272 groups of 8 wires, regardless
  of whether there were any hits in that group.  A simple O(n) check now avoids
  many of these O(n^2) searches.  (Keep in mind that the real pattern takes
  O(8^levels) to be found, so O(280^2) is still quite small for multiplicity
  per plane around 1 track.)



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@679 b9700a98-9937-0410-b777-f318d6eb5bff