I added discrete Fourier transform functionality to the GUI, for the main detectors. So we can now display histograms, time graphs, and DFTs.
I wanted to test this and therefore had to add the
possibility of a sinusoidal drift to the mock data
generator, which in turn meant that I had to edit
classes in the Analysis and Parity directories.

The following changes were made:

QwIntegrationPMT:

	Added a function: 
	      SetRandomEventDriftParameters
	     
	Which sets the amplitude, phase, and frequency for
        the sinusoidal "drift". This follows the spirit in
        which the other mock random event parameters are set.

	Added a function:
	      SetEventNumber

	I couldn't find a way to get the actual event number 
	from the mock data generator class into the channel
        class, so I just made a new setter function. However
	this may well be a duplication of something that
	already worked in some other way ... Can someone please 
        correct this if that is the case.

QwVQWK_Channel

	Added the same function as above: 
	      SetRandomEventDriftParameters

	Added the same function as above:
	      SetEventNumber	      

	Modified the function:
 	      RandomizeEventData

Modified QwMockDataGenerator.cc to support the
new functionality.

All in all these were minor modifications.

I am still a little mystified about the relationship
between events that I specify when I run the mock
data generator (e.g. 100000) and the corresponding
tree entries I get when I run the analyzer on the
output file (1086). 

I know I could sift through the code to find out,
but if some could enlighten me and answer the following
questions, that would be really nice (Wouter ?).

What is it that I specify when I run 100000 events 
with the mock data generator ?

What is the timing? I think I know that each tree entry 
that is in the analyzer output (mps ?) corresponds to
1.06 ms of data, which consists of four blocks of 
integrated data, but I am unclear as to what the event
specification for the mock data generator means.

Thanks,

Michael
        



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@693 b9700a98-9937-0410-b777-f318d6eb5bff