This is an experimental branch trying to get multithreading to work in the analyzer (tracking analyzer mainly).  It is based on the QwControl class, and should theoretically allow to have many analyzer objects at the same time.  Right now it just has one analyzer object that is spun off as a thread for every CODA event.  This is patently unhealthy, which becomes clear soon after you start the executable.  Here be dragons...
Architecture:
The basic class is the VQwSystem (confusing, I know, this is not a sub-
system nor do subsystems depend on this, but there are only few abstract
names to choose from).  The VQwDataserver and VQwAnalyzer inherit from
the VQwSystem.  They provide the virtual interface to the data (basically
a wrapper around the QwEventBuffer that handles the CODA stream and file
opening) and the virtual interface to the analysis (this is different for
tracking and parity).

The QwRoot class provides the frame of a general analysis: do some init,
then go into a data loop.  It is a thread (that lives through the entire
program's life).  It spins threads for the analyzers (which live for one
event only; this is what might make it quite useless due to the cost of
spinning the thread).

The QwControl class is just a TRint ROOT-like interface, which might not
be so useful after all, but comes for free in a different thread.  While
the analysis is running, you can start a TBrowser and inspect the histos.

The implementation classes QwTrackingDataserver and QwTrackingAnalyzer do
the specific things for tracking: get the hitlist, process it through the
tracking worker, write to histos and trees.  The analyzer knows about the
subsystem array, the dataserver only knows about the CODA stream/file.


I don't know where this branch is going to go, but it could be a proof of
concept for the multithreaded analysis, which we could choose to implement
in a less convoluted way later.  Or it might not work in the way I hope it
will...


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.multithread@722 b9700a98-9937-0410-b777-f318d6eb5bff