Note: This is branch pking.f1tdc_decoder_change.
Modified the MQwF1TDC::SubtractReference method to consistently
decode both R2 and R3 data.
The R3 event processing is currently dependent on two somewhat
unusual behaviors of MQwF1TDC::SubtractReference:
	-  An additional time-shift is added to the "real time";
	   in some future revision, this would be better done
	   by a different offset correction function (which
	   would be specific to the VDC class and not built
	   into the F1TDC decoder).
	-  The "real time" is negative compared to the 
	   difference "raw - reference", which is what you might
	   expect a SubtractReference method to provide.
The MQwF1TDC class has parameters in the constructor and a setter
method to set these parameters.

  Eventually, the "fTimeShift" and "fSignFlip" internal parameters
should be removed from the MQwF1TDC class, and their purpose should
be filled by appropriate functions in the QwDriftChamberVDC class.
The QwDriftChamberHDC class expects "fTimeShift" to be zero and
"fSignFlip" to be false.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.f1tdc_decoder_change@726 b9700a98-9937-0410-b777-f318d6eb5bff