Wrote doxygen documentation for the tree search and tree generation algorithms.
The tree search is fully documented for region 3; region 2 is similar, and there
is a general (unused) code block that gives the generalized approach.  I added
some more appropriately named variables, and created references to fNumLayers
(= number of layers in the tree) that indicate what the layers represent in each
region.

The tree generation is not fully documented, but I started and will continue
writing the docs here or there.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@749 b9700a98-9937-0410-b777-f318d6eb5bff