This commit includes modification to the options parsing routines.  There were some problems with the try/catch logic used when parsing configuration files and command line options that caused valid options to be trown away.  This should improve that. Only the mock data generator code was not updated because I don't want to give away the parameters I used in the latest version of mock data :-)
Also, this commit removes the need for QwCommandLine in all source files that are in
the repository.  The calling arguments are unchanged: -r run1:run2 -e event1:event2.

Defining command line and config file options for subsystems was changed slightly.
Now you need to define the static method DefineOptions(QwOptions& options) and set the
options on the specified argument.  Do not use the global gQwOptions anymore.

Note: this can probably be merged into the spayde branch to fix the problem with the
options for setting the database connection info.  In QwDatabase the function DefineOptions
will have to be modified.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@763 b9700a98-9937-0410-b777-f318d6eb5bff