Merged from the trunk to pick up recent changes.  Merged revisions 620:770.
Had to resolve conflicts in GNUMakefile (added database executables) and in QwColors.h (no idea why).

Cleaned up code so merged version would compile and run.  This involved updating DefineOption() calls in QwDatabase to come into agreement with updated QwOptions class.

This merge introduced a segmentation violation on my test program, qwdb_test.  I tracked it down to the debugging line in the QwDatabase destructor.  When I comment out the line output on QwDebug the segfault goes away.  This is sort of mystifying to me right now but not a deal-killer.  I commented out the debugging line for right now.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/spayde@773 b9700a98-9937-0410-b777-f318d6eb5bff