QwHelicity updates
Random seed size is either 24 bit or 30 bit with default is 24 bit. This can be set in the map file

	RandSeedBits=24
	or
	RandSeedBits=30

Need to set phase number offset to 1 (when pattern number starts with 1) or 0 (when pattern number starts with 0) inside the map file. 
Default value is set to 1 within the code.

QwHelicityPattern class update
Implemented a mechanism to identify if the pattern-phase-number starts with 0 or 1
 
Now analyzer can decode quartets and octets.

Rakitha



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@790 b9700a98-9937-0410-b777-f318d6eb5bff