Implemented the publishing of variables for intersubsystem communication.
Basically, the subsytems can now implement
  virtual const Bool_t PublishInternalValues() const
by successive calls to
  const Bool_t PublishInternalValue(const TString name, const TString desc) const
For example, in the QwBeamLine subsystem we have
  PublishInternalValue("q_targ", "Calculated charge on target")

You will get an error if the variable name already exists in the subsystem array.

You can retrieve published variables from other subsystems by calling
  if (RequestExternalValue(variable->GetElementName(), variable))
as for example in the QwMainCerenkovDetector.  If the variable does not exist or
is not published, you will get an error.  At the moment, the beam line only publishes
the charge q_targ, but the main detector tries to get both the charge and the
positions x_targ and y_targ on target.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.intersubsystem_communication@796 b9700a98-9937-0410-b777-f318d6eb5bff