Added a few lines of documentation to make clear why the DefineOptions method of the QwLog class takes a pointer as opposed to a reference.
This is needed to avoid including the QwOptions header file in the
QwLog header file.  This is the only way out of the conundrum that
is caused by the templated QwOptions::GetValue<>() function.  That
function prints an error if the value is not defined and thus needs
to know about the QwLog objects by including the QwLog header file
in the QwOptions header file.  Therefore we cannot include the QwOptions
header file in QwLog, or we end up with a circular include...

All other subsystems use references in DefineOptions(QwOptions &) !!!


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@819 b9700a98-9937-0410-b777-f318d6eb5bff