Moved the run and event range option flag handling out of the QwOptions class and into the QwEventBuffer class.
The QwEventBuffer is now directly responsible for 
maintaining the loops through runs and events.
An option flag "chainfiles" allows file segments to
be chained together to form one full analysis; this
option is currently enabled by default.

Changed the run and event loops of the QwTracking,
QwAnalysis_ADC, and QwAnalysis_BeamLine main
analyzers to use the new QwEventBuffer methods.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@821 b9700a98-9937-0410-b777-f318d6eb5bff