New class which performs interpolation on grids.  It is a template class, so you can use it to interpolate TVector3 objects (which might be faster than using three QwInterpolator objects).  It works with an arbitrary number of grid dimensions (set on construction), and does multi-linear interpolation transparently when calling GetValue(coord[]).  For reasons of speed, the indices are not always checked when you call index getters directly, although you shouldn't do that anyway.  The index getters are not private because currently the grid is filled from the outside.
This will be used in several parts of the tracking code that use inter-
polation already: magnetic field map, tracking lookup table, and the
interpolation inside the lookup table methods.  This should remove the
dependency on ROOT mathmore when fully implemented (soon).

Routines to read and write grids from ROOT file (trajectory lookup table),
text file (magnetic field grid), and binary file (newer, faster, smaller
magnetic field grid) are in progress.

Magnetic field interpolation is the main reason why tracking is slow,
so the readability of the code was of secondary importance... sorry.
Documentation is still there, of course.  The interpolation is done in
5 lines, and a TVector3 should hopefully increase the speed even more.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.restruct-Q2-det@837 b9700a98-9937-0410-b777-f318d6eb5bff