Modified the QwCombinedBPM and QwCombinedBCM classes to improve the time taken for calculations. This involved in creating a new function for the QwVQWK channel in order to get the product of two channels.
In the last update I did, the combined BPM was able to calculate the position and angle on the target. In this I changed that to be the position and slope instead.

Fixed a bug in the LoadGeometry () function in QwBeamLine.cc that was causing the analyzer to crash when an unknown device was read in. Now, the unknown device is skipped.

If anyone notice any errors caused by the combined beamline device classes, please let me know!



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@839 b9700a98-9937-0410-b777-f318d6eb5bff